API service updated
