+++ gi_service/contract/rest/person_api_contract.yaml
@@ -159,49 +159,6 @@ paths:
               example:
                 detail: "An unexpected error occurred"
 
-  /v1/person/all-presidents:
-    get:
-      tags:
-        - Person
-      summary: Get all presidents.
-      description: Returns a list of all presidents with their terms and gazettes published during those terms.
-      operationId: getAllPresidents
-      responses:
-        "200":
-          description: Successfully returned all presidents
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/AllPresidentsResponse"
-              example:
-                presidents:
-                  - id: "pres-1"
-                    name: "John Doe"
-                    terms:
-                      - start: "2010-01-01"
-                        end: "2015-01-01"
-                        gazettes_published:
-                          - date: "2012-05-15"
-                            ids: ["gzt-101", "gzt-102"]
-                  - id: "pres-2"
-                    name: "Jane Smith"
-                    terms:
-                      - start: "2015-01-01"
-                        end: null
-                        gazettes_published: []
-
-        "500":
-          description: Server Error
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  detail:
-                    type: string
-              example:
-                detail: "An unexpected error occurred"
-
 components:
   schemas:
     MinistryHistoryItem:
@@ -313,76 +270,3 @@ components:
           type: integer
           description: "The person's age"
           examples: [45]
-
-    AllPresidentsResponse:
-      type: object
-      required:
-        - presidents
-      properties:
-        presidents:
-          type: array
-          description: "List of all presidents sorted by their latest term date"
-          items:
-            $ref: "#/components/schemas/PresidentInfo"
-
-    PresidentInfo:
-      type: object
-      required:
-        - id
-        - name
-        - terms
-      properties:
-        id:
-          type: string
-          description: "Unique president identifier"
-          example: "pres-1"
-        name:
-          type: string
-          description: "Full name of the president"
-          example: "John Doe"
-        terms:
-          type: array
-          description: "List of presidential terms, sorted by latest term first"
-          items:
-            $ref: "#/components/schemas/PresidentTerm"
-
-    PresidentTerm:
-      type: object
-      required:
-        - start
-        - end
-        - gazettes_published
-      properties:
-        start:
-          type: string
-          format: date
-          description: "Term start date (YYYY-MM-DD)"
-          example: "2010-01-01"
-        end:
-          type: string
-          format: date
-          nullable: true
-          description: "Term end date (YYYY-MM-DD), null if ongoing"
-          examples: ["2015-01-01", null]
-        gazettes_published:
-          type: array
-          items:
-            $ref: "#/components/schemas/GazettePublished"
-
-    GazettePublished:
-      type: object
-      required:
-        - date
-        - ids
-      properties:
-        date:
-          type: string
-          format: date
-          description: "Date the gazettes were published"
-          examples: ["2012-05-15"]
-        ids:
-          type: array
-          items:
-            type: string
-          description: "List of gazette IDs published on this date"
-          example: ["gzt-101", "gzt-102"]
